+++ resources/views/users/show.blade.php
@@ -96,7 +96,7 @@
                     @endcan
                 </dt>
                 <dd x-bind:class="{ 'user-edit-active': editRole }">
-                    <span x-show="!editRole">{{ $user->role->translation() }}</span>
+                    <span x-show="!editRole">{!! formatUserRoleLabel($user->role) !!}</span>
                     <x-ig::form class="editable-skip" :recaptcha="false" x-show="editRole" :action="route('users.update', $user)">
                         <div class="input-group">
                             <select name="role" class="form-select" value="{{ $user->role->value }}">